chore(publish): split env into client and schema
