+++ packages/nextjs-template/utils/validation.js
@@ -1,7 +1,7 @@
 const fs = require("fs-extra");
 const path = require("path");
 const { parse, treeMenuSchema } = require("@dendronhq/common-all");
-const { env } = require('./env')
+const { env } = require('../env/server')
 
 const dataDir = env.DATA_DIR
 if (!dataDir) {
